소수를 입력받을때 프로그램의 비정상적 종료현상을 해결하기위한 코드 추가
키,몸무게, 희망몸무게를 정수가 아닌 소수를 입력받았을때 비정상적 종료를 해결하기위해 에러처리 추가 및 int형으로 바꾸었던 몇몇 변수들을 int형이아닌 float형으로 바꾸도록하여 해결하였습니다

+++ toModule2/calcDB.py
@@ -22,7 +22,7 @@ class CalcDB:
         exercise_minutes = whole_exercise_minutes / int(exercise_day)  # 하루에 운동해야 하는 시간
         return [randexer,exercise_minutes]
     def calcBMI(self,weight,height):
-        BMI = int(weight) / (int(height) / 100) ** 2
+        BMI = float(weight) / (float(height) / 100) ** 2
         if BMI < 20:
             condition = '저체중'
         elif 20 < BMI < 24:
